Cleaning up signatures
